Using FunctionList refactor BaseFunction.
