+++ paddle/function/CMakeLists.txt
@@ -3,6 +3,8 @@ file(GLOB cpp_files . *Op.cpp)
 
 list(APPEND h_files Function.h)
 list(APPEND cpp_files Function.cpp)
+list(APPEND h_files FunctionList.h FuncConfig.h BufferArgs.h)
+list(APPEND cpp_files FunctionList.cpp)
 list(APPEND cpp_files BufferArg.cpp)
 
 if(WITH_GPU)
