+++ paddle/gserver/layers/ContextProjection.h
@@ -15,7 +15,6 @@ limitations under the License. */
 #pragma once
 
 #include "Projection.h"
-#include "WithFunction.h"
 namespace paddle {
 
 /**
@@ -41,7 +40,7 @@ namespace paddle {
  *
  * The config file api is context_projection.
  */
-class ContextProjection : public Projection, public WithFunction {
+class ContextProjection : public Projection {
 public:
   /**
    * Constructor. If context_start is zero and context_lenth is one, it will
