+++ paddle/gserver/layers/CosSimLayer.h
@@ -14,7 +14,6 @@ limitations under the License. */
 
 #pragma once
 #include "Layer.h"
-#include "WithFunction.h"
 #include "paddle/math/Matrix.h"
 #include "paddle/utils/ThreadLocal.h"
 
@@ -32,7 +31,7 @@ namespace paddle {
  *
  * The config file api is cos_sim.
  */
-class CosSimLayer : public Layer, public WithFunction {
+class CosSimLayer : public Layer {
 public:
   explicit CosSimLayer(const LayerConfig& config) : Layer(config) {}
